Disregard "Auto Newline" user setting and always assume auto newline in FileReader
This may not be the proper place to fix this but it works nicely at least as a temporary measure. The alternative is a potential DoS when trying to open large files as discussed in #352.

Now FileReader always splits on newlines regardless of what the user selected for "Auto Newline" option from the menu. Fixes #352.