Fix restoring position of maximized and minimized window too
If the window was closed in the maximized (or, more rarely, but still
possible, minimized) state, restore it in the same state during the next
run.

Take care to use the appropriate show command for the ShowWindow() call
done in Main(), as using SW_SHOWMAXIMIZED with SetWindowPlacement()
doesn't work quite as expected.