Make it possible to override NUGET_PACKAGES_PATH on command line
When PackagesRoot is set to some global directory, nuget packages are
installed in it and not under application-specific nuget directory, so
using the default value of NUGET_PACKAGES_PATH didn't work and CMake
configure step failed because it didn't find third-party packages files.

Fix this by allowing to specify -DNUGET_PACKAGES_PATH=$PackagesRoot on
CMake command line to override the default value.